バス const reg の裸数値初期化(ニブル分解)を実装する
`const reg[W] n = <literal>;` を、値を4bitニブルとして下位レーンから分解し
各レーンに独立したスカラ const を割り当てる形で追加する(issue #140)。
monitor のバス→整数パッキング(§7.4.1)をちょうど逆にした対応にすることで、
書いたリテラルが monitor の %x でそのまま読み返せる対称性を持たせた。
ニブル(0-15)が信号強度の値域と一致するため、比較器の閾値のようにレーン
ごとに異なる中間強度を持つ定数を1個のリテラルで表せる(bus_reg_side.rv の
SUB4 で毎tick var 経由で与えていたものが定数化できる)。

設計合意では issue が挙げた「ブロードキャスト」(既存の fan-out と等価で
新規性がない)・「1bitごとの0/15分解」(閾値のような0-15中間値を表現できず
monitor のパッキングとも非対称)を退け、ニブル分解を採用した。

parser 側は tok の有無で「素子代入」か「裸数値」かを判定するだけに単純化し、
qual との組み合わせ妥当性は interp.rs 側に集約した。裸数値初期化の判定に
strength>=0 を「数値の有無」の代用にしていた箇所は、width-16 ニブル分解で
上位ビットが立つと符号反転して誤判定するため、has_num フラグへ切り出した。
同じ理由で lexer の16進/2進リテラル読み取りも wrapping 演算に変更し、
0xffffffffffffffff のような64bitぎりぎりの値でパニックしないようにした。

+++ src/ast.rs
@@ -40,10 +40,13 @@ pub enum WidthExpr {
 }
 
 /// reg 宣言の初期化子。`strength == -1` は信号強度未指定。
-/// `tok == None` は素子トークンを伴わない裸数値初期化(`const reg n = 15;`)。
+/// `tok == None` は素子トークンを伴わない裸数値初期化(`const reg n = 15;`、
+/// バスなら `const reg[W] n = 0x3808;` のニブル分解、issue #140)。
+/// `i64` なのは、バス幅 16(64 bit)までの裸数値をそのまま保持するため
+/// (スカラは 0-15 の範囲チェックが別途入るので i32 で足りるが、フィールドは共有)。
 #[derive(Debug, Clone)]
 pub struct RegInit {
-    pub strength: i32,
+    pub strength: i64,
     pub tok: Option<String>,
 }
